+++ examples/ataxx.rs
@@ -1,7 +1,6 @@
 use bullet_lib::{
-    default::{loader, outputs, Loss, TrainerBuilder},
+    default::{inputs::SparseInputType, loader, outputs, Loss, TrainerBuilder},
     format::AtaxxBoard,
-    inputs::SparseInputType,
     lr, optimiser, wdl, Activation, LocalSettings, TrainingSchedule, TrainingSteps,
 };
