+++ src/trainer/default/loader/sfbinpack.rs
@@ -5,7 +5,7 @@ use sfbinpack::{
     CompressedTrainingDataEntryReader, TrainingDataEntry,
 };
 
-use crate::{format::ChessBoard, loader::DataLoader, rng::SimpleRand};
+use crate::{default::loader::DataLoader, format::ChessBoard, rng::SimpleRand};
 
 fn convert_to_bulletformat(entry: &TrainingDataEntry) -> ChessBoard {
     let mut bbs = [0; 8];
